Refund Invoice Pdf

+++ internal/logic/metric_event/metric_event.go
@@ -37,7 +37,7 @@ func NewMerchantMetricEvent(ctx context.Context, req *MerchantMetricEventInterna
 	utility.Assert(met != nil, "metric not found")
 	utility.Assert(met.MerchantId == req.MerchantId, "code not match")
 	// check the only paid subscription
-	sub := query.GetLatestActiveOrIncompleteSubscriptionByUserId(ctx, int64(user.Id), req.MerchantId)
+	sub := query.GetLatestActiveOrIncompleteSubscriptionByUserId(ctx, user.Id, req.MerchantId)
 	utility.Assert(sub != nil, "user has no active subscription")
 
 	// property determine
